requirements.txt
allure_python_commons==2.13.5
Faker==28.1.0
pytest==8.3.2
selenium==4.24.0
